backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
torch==2.1.0
zennit==0.5.1
biopython==1.81
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
schedule==1.2.1